Also remove TormaystarkasteluTormaysService interface
